DEV-1023 Migrate contactDetail client component to server component
