+++ src/app/crm/(pages)/contacts/[id]/page.tsx
@@ -1,91 +1,19 @@
-'use client';
 import React from 'react';
-import {
-  Tab,
-  TabPanel,
-  Tabs,
-  TabsBody,
-  TabsHeader,
-} from '@material-tailwind/react';
-import ContactService from '@services/ContactService';
-import useAsync from '@utils/useAsync';
-import MainLayoutCRM from 'app/crm/components/layout/MainLayoutCRM';
-import { SvgCalendar, SvgMedicine, SvgPhone } from 'app/icons/Icons';
-import { Flex } from 'designSystem/Layouts/Layouts';
+import useAsyncServer from '@utils/useAsyncServer';
+import ContactService from 'app/crm/services/ContactService';
+
+import ContactDetailPageBase from './ContactDetailPageBase';
 
 interface ContactDetailProps {
   params: { id: string };
 }
 
 export default function ContactDetailPage({ params }: ContactDetailProps) {
-  const {
-    data: contactDetail,
-    isLoading,
-    error,
-  } = useAsync<any>(ContactService.ContactDetail(params.id));
-
-  const data = [
-    {
-      label: 'Tareas',
-      value: 'Tareas',
-      icon: <SvgPhone />,
-      desc: 'Test',
-    },
-    {
-      label: 'Datos personales',
-      value: 'Datos personales',
-      icon: <SvgMedicine />,
-      desc: 'Test1',
-    },
-    {
-      label: 'Comentarios',
-      value: 'Comentarios',
-      icon: <SvgPhone />,
-      desc: 'Test2',
-    },
-    {
-      label: 'Llamadas',
-      value: 'Llamadas',
-      icon: <SvgPhone />,
-      desc: 'Test3',
-    },
-    {
-      label: 'Citas',
-      value: 'Citas',
-      icon: <SvgCalendar />,
-      desc: 'Test4',
-    },
-    {
-      label: 'Presupuestos',
-      value: 'Presupuestos',
-      icon: <SvgPhone />,
-      desc: 'Test5',
-    },
-  ];
+  const contactDetail = useAsyncServer(ContactService.ContactDetail(params.id));
 
   return (
-    <MainLayoutCRM>
-      <div className="rounded-xl bg-white ml-64 mt-2 mr-4 h-screen">
-        <Tabs value="Tareas">
-          <TabsHeader>
-            {data.map(({ label, value, icon, desc }) => (
-              <Tab key={value} value={value}>
-                <Flex layout="row-center" className="">
-                  {icon}
-                  {label}
-                </Flex>
-              </Tab>
-            ))}
-          </TabsHeader>
-          <TabsBody>
-            {data.map(({ value, desc }) => (
-              <TabPanel key={value} value={value}>
-                {desc}
-              </TabPanel>
-            ))}
-          </TabsBody>
-        </Tabs>
-      </div>
-    </MainLayoutCRM>
+    <>
+      <ContactDetailPageBase contactDetail={contactDetail} />
+    </>
   );
 }
